update: 替换早期整体测试代码为正常业务代码 update: 数据库int添加兼容套层

+++ database/perm.go
@@ -47,7 +47,7 @@ func SetPermission(PermLevel int, OpenID string) (isSusses bool, err error) {
 	UpdatePk := new(tablestore.PrimaryKey)
 	UpdatePk.AddPrimaryKeyColumn("OpenID", OpenID)
 	UpdateRowChange.PrimaryKey = UpdatePk
-	UpdateRowChange.PutColumn("permission", PermLevel)
+	UpdateRowChange.PutColumn("permission", int64(PermLevel))
 	UpdateRowRequest.UpdateRowChange = UpdateRowChange
 	_, err = client.UpdateRow(UpdateRowRequest)
 	if err != nil {
